Some refinements

+++ cap/src/agent.py
@@ -2,6 +2,7 @@ from robot import RobotArm
 import numpy as np
 import copy
 import os
+import utility
 
 
 class RobotArmAgent(object):
@@ -21,11 +22,14 @@ class RobotArmAgent(object):
         y_range = np.arange(dim[1][0], dim[1][1], self.tolerance)
         z_range = np.arange(dim[2][0], dim[2][1], self.tolerance)
 
-        self.action_type1 = 'move_claw'
-        self.action_type2 = 'engage_claw'
+        self.action_type1 = 'move_gripper'
+        self.action_type2 = 'engage_gripper'
         self.invalid_state = [-100, -100, -100, False]
-        self.actions = []
 
+        # Actions consist of
+        # a) Gripper Enable/Disable
+        # b) Goto location (x, y, z)
+        self.actions = []
         self.actions.append([self.action_type2, True])
         self.actions.append([self.action_type2, False])
 
@@ -36,9 +40,7 @@ class RobotArmAgent(object):
 
         self.total_actions = len(self.actions)
 
-        # The states the environment can be in, consists of:
-        # a. Position of the cylinder (the item to pick)
-
+        # The states the environment can be in, is made up of all the locations and if gripper is engaged/disengaged
         self.states = []
         for x in x_range:
             for y in y_range:
@@ -48,6 +50,7 @@ class RobotArmAgent(object):
 
         self.states.append(self.invalid_state)  # invalid state, the last state.
         self.total_states = len(self.states)
+        self.invalid_states_index = self.total_states - 1
 
         print("There are {0} actions.".format(self.total_actions))
         print("There are {0} states.".format(self.total_states))
@@ -56,8 +59,12 @@ class RobotArmAgent(object):
 
         # Give the claw engaging actions high initial values
         for index, val in enumerate(self.q_table):
-            self.q_table[index][0] = 2
-            self.q_table[index][1] = 2
+            if not self.states[index][3]:
+                self.q_table[index][0] = 2
+                self.q_table[index][1] = -1
+            else:
+                self.q_table[index][0] = -1
+                self.q_table[index][1] = 2
 
         self.state_id = None
         self.actionstate_prev = {}
@@ -65,15 +72,14 @@ class RobotArmAgent(object):
 
     def load_qtable(self):
         f = 'qtables/qtable.txt.npy'
-        self.q_table = np.load(f)
+        if os.path.exists(f):
+            self.q_table = np.load(f)
 
     def save_qtable(self):
         print('Saving Q Table')
         f = 'qtables/qtable.txt.npy'
-        try:
+        if os.path.exists(f):
             os.remove(f)
-        except:
-            pass
         np.save(f, self.q_table)
 
     def init(self):
@@ -82,7 +88,7 @@ class RobotArmAgent(object):
         self.actionstate_prev = {}
         self.actionstate_curr = {}
 
-    def get_canonical_position(self, handle, claw_enabled):
+    def get_canonical_position(self, handle, gripper_enabled):
         pos = self.robot.get_position(handle)
 
         state_id = 0
@@ -90,11 +96,11 @@ class RobotArmAgent(object):
             if abs(state[0] - pos[0]) <= self.tolerance \
                     and abs(state[1] - pos[1]) <= self.tolerance \
                     and abs(state[2] - pos[2]) <= self.tolerance \
-                    and state[3] == claw_enabled:
+                    and state[3] == gripper_enabled:
                 return state_id
             state_id += 1
 
-        return self.total_states - 1  # The last state, which is the invalid state, will get returned
+        return self.invalid_states_index  # The last state, which is the invalid state, will get returned
 
     def update_actionstate(self, action_id):
         self.actionstate_curr['action_id'] = action_id
@@ -102,12 +108,11 @@ class RobotArmAgent(object):
         self.actionstate_curr['state'] = self.states[self.actionstate_curr['state_id']]
         self.actionstate_curr['cylinder_position'] = self.robot.get_position(self.robot.cylinder_handle)
         self.actionstate_curr['bin_position'] = self.robot.get_position(self.robot.bin_handle)
-        self.actionstate_curr['claw_position'] = self.robot.get_position(self.robot.claw_handle)
-
-        self.actionstate_curr['claw_cylinder_distance'] = self.distance(self.actionstate_curr['claw_position'],
-                                                                        self.actionstate_curr['cylinder_position'])
-        self.actionstate_curr['bin_cylinder_distance'] = self.distance(self.actionstate_curr['bin_position'],
-                                                                       self.actionstate_curr['cylinder_position'])
+        self.actionstate_curr['claw_position'] = self.robot.get_position(self.robot.gripper_handle)
+        self.actionstate_curr['claw_cylinder_distance'] = utility.distance(self.actionstate_curr['claw_position'],
+                                                                           self.actionstate_curr['cylinder_position'])
+        self.actionstate_curr['bin_cylinder_distance'] = utility.distance(self.actionstate_curr['bin_position'],
+                                                                          self.actionstate_curr['cylinder_position'])
         self.actionstate_curr['cylinder_in_bin'] = self.robot.is_object_in_bin()
         self.actionstate_curr['is_cylinder_held'] = self.robot.is_object_held()
 
@@ -126,46 +131,40 @@ class RobotArmAgent(object):
             self.robot.goto_position(action[1])
         elif action[0] == self.action_type2:
             print('Action: Engaging/Disengaging claw')
-            self.robot.enable_claw(action[1])
+            self.robot.enable_grip(action[1])
 
-    @staticmethod
-    def distance(pos1, pos2):
-        x2 = np.square(pos1[0] - pos2[0])
-        y2 = np.square(pos1[1] - pos2[1])
-        z2 = np.square(pos1[2] - pos2[2])
-
-        return np.sqrt(x2 + y2 + z2)
-
-    def update_q(self, state, action, reward, state_new):
-        print(state, action, reward, state_new)
+    def update_q_table(self, state, action, reward, state_new):
+        print("Reward: ", reward)
         q_sa = self.q_table[state, action]
         td_error = reward + self.gamma * np.max(self.q_table[state_new]) - q_sa
         self.q_table[state, action] = q_sa + self.alpha * td_error
 
     def observe_state(self):
-        return self.get_canonical_position(self.robot.center_handle, self.robot.claw_enabled)
+        return self.get_canonical_position(self.robot.gripper_handle, self.robot.gripper_enabled)
 
     def calculate_reward(self):
 
         state_new = self.actionstate_curr['state']
 
         terminate = False
+        is_pass = False
+
         if state_new == self.invalid_state:
             reward = -10
             print('Penalty: Reached invalid state, terminating')
             terminate = True
-        elif (ra.get_position(ra.cylinder_handle)[2] - ra.cylinder_z_locus) < -0.005:
+        elif (ra.get_position(ra.cylinder_handle)[2] - ra.cylinder_z_locus) < (-1 * self.tolerance):
             reward = -10
             print('Penalty: Cylinder has fallen, terminating')
             terminate = True
-        elif self.distance(ra.bin_position, self.actionstate_curr['bin_position']) > 0.005:
+        elif utility.distance(ra.bin_position, self.actionstate_curr['bin_position']) > self.tolerance:
             reward = -10
             print('Penalty: Bin has shifted, terminating')
             terminate = True
         elif len(self.actionstate_prev) > 0 and self.actionstate_curr['state'] == self.actionstate_prev['state']:
             reward = -1
             print('Penalty: Previous and current state is same')
-        elif self.robot.claw_enabled \
+        elif self.robot.gripper_enabled \
                 and not self.actionstate_curr['is_cylinder_held'] \
                 and not self.actionstate_curr['cylinder_in_bin']:
             reward = -1
@@ -174,32 +173,29 @@ class RobotArmAgent(object):
                 and self.actionstate_prev['is_cylinder_held'] \
                 and not self.actionstate_curr['is_cylinder_held'] \
                 and not self.actionstate_curr['cylinder_in_bin']:
-            reward = -2
+            reward = -1
             print('Penalty: Claw did not drop the cylinder in the bin')
-        elif self.robot.claw_enabled \
+        elif self.robot.gripper_enabled \
                 and self.actionstate_curr['is_cylinder_held'] \
                 and not self.actionstate_curr['cylinder_in_bin']:
-            if (ra.get_position(ra.cylinder_handle)[2] - ra.cylinder_z_locus) > 0.001:
+            if (ra.get_position(ra.cylinder_handle)[2] - ra.cylinder_z_locus) > self.tolerance:
                 reward = 10
-                print('Reward: Claw could grab and lift the cylinder !!!!!!!!!!!!!!!!!!!!!!!!!')
+                print('Reward: Claw could grab *and lift* the cylinder !!!!!!!!!!!!!!!!!!!!!!!!!')
             else:
                 reward = 5
                 print('Reward: Claw could grab the cylinder !!!!!!!!!!!!!!!!!!!!!!!!!')
-        elif not self.robot.claw_enabled \
-                and not self.actionstate_curr['is_cylinder_held'] \
-                and self.actionstate_curr['cylinder_in_bin']:
+        elif self.actionstate_curr['cylinder_in_bin']:
             reward = 100
             print('Reward: Cylinder in bucket !!!!!!!!!!!!!!!!!!!!!!!!!')
             terminate = True
+            is_pass = True
         else:
-            # post_distance = self.distance(state_new, claw_position_new)
-            # forward = pre_distance - post_distance
-            d1 = self.distance(state_new, self.actionstate_curr['bin_position'])
-            d2 = self.distance(state_new, self.actionstate_curr['claw_position'])
-            reward = 5 - (d1 + d2)
+            #d1 = utility.distance(state_new, self.actionstate_curr['bin_position'])
+            #d2 = utility.distance(state_new, self.actionstate_curr['claw_position'])
+            #reward = 5 - (d1 + d2)
             reward = -1
 
-        return reward, terminate
+        return reward, terminate, is_pass
 
     def step_through(self):
         self.actionstate_prev = copy.deepcopy(self.actionstate_curr)
@@ -208,46 +204,38 @@ class RobotArmAgent(object):
         self.update_actionstate(action_id)
         state_id_new = self.actionstate_curr['state_id']
 
-        reward, terminate = self.calculate_reward()
+        reward, terminate, is_pass = self.calculate_reward()
 
         # I was in this state, I took this action, I got this reward, and I reached the new state
-        self.update_q(self.state_id, action_id, reward, state_id_new)
+        self.update_q_table(self.state_id, action_id, reward, state_id_new)
         self.state_id = state_id_new
-        return terminate
+        return terminate, is_pass, reward
 
 
 ra = RobotArm('127.0.0.1', 19997)
-raa = RobotArmAgent(ra, epsilon=0)
-
-
-"""
-raa.init()
-state = raa.states[raa.observe_state()]
-print(state[0])
-print(state[1])
-print(state[2])
-print(state)
-print(ra.get_position(ra.cylinder_handle))
-
-
-ra.goto_position(state)
-ra.enable_claw(True)
-ra.goto_position([-0.3, -0.11, 0.12])
-
-
-"""
-
-episodes = 1000
+raa = RobotArmAgent(ra)
+episodes = 10000
+episode_num = 0
 raa.load_qtable()
+
 while episodes > 0:
     print('=============================================> Episode ', episodes)
     raa.init()
     i = 0
+    is_pass = False
+    total_reward = 0
+    episode_num += 1
     while i < 100:
-        terminate1 = raa.step_through()
-        if terminate1:
+        terminate_x, is_pass_x, reward = raa.step_through()
+        total_reward += reward
+        if terminate_x:
+            is_pass = is_pass_x
             break
         i += 1
 
+    # save episode_id, is_pass, total_reward, (i+1)
+    stat_file = open('qtables/1aug.txt', 'a')
+    stat_file.write("{0}, {1}, {2}, {3}\n".format(episode_num, is_pass, total_reward, i+1))
+    stat_file.close()
     raa.save_qtable()
     episodes -= 1
